requirements.txt
streamlit
pandas
openpyxl
anthropic
python-dotenv
psycopg2-binary
